+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Issue with multiple detected error popups showing after launch [#642]
+- Don't show Java setting when using Java provided by Minecraft (match Instance settings)
 
 ### Misc